Move all the tasks to the root of the repository
