+++ tasks/e2e_framework/__init__.py
@@ -1,0 +1,1 @@
+# mypy: disable-error-code="arg-type"
