+++ tasks/e2e_framework/azure/__init__.py
@@ -1,7 +1,7 @@
 from invoke.collection import Collection
 
-from tasks.azure.aks import create_aks, destroy_aks
-from tasks.azure.vm import create_vm, destroy_vm
+from tasks.e2e_framework.azure.aks import create_aks, destroy_aks
+from tasks.e2e_framework.azure.vm import create_vm, destroy_vm
 
 collection = Collection()
 collection.add_task(destroy_vm)
