+++ tasks/e2e_framework/azure/common.py
