+++ tasks/e2e_framework/gcp/__init__.py
@@ -1,8 +1,8 @@
 from invoke.collection import Collection
 
-from tasks.gcp.gke import create_gke, destroy_gke
-from tasks.gcp.openshift import create_openshift, destroy_openshift
-from tasks.gcp.vm import create_vm, destroy_vm
+from tasks.e2e_framework.gcp.gke import create_gke, destroy_gke
+from tasks.e2e_framework.gcp.openshift import create_openshift, destroy_openshift
+from tasks.e2e_framework.gcp.vm import create_vm, destroy_vm
 
 collection = Collection()
 collection.add_task(destroy_vm)
