+++ tasks/e2e_framework/gcp/common.py
