+++ tasks/e2e_framework/localpodman/__init__.py
@@ -1,6 +1,6 @@
 from invoke.collection import Collection
 
-from tasks.localpodman.vm import create_vm, destroy_vm
+from tasks.e2e_framework.localpodman.vm import create_vm, destroy_vm
 
 collection = Collection()
 collection.add_task(destroy_vm)
